app/src/main/java/com/jaspersoft/android/jaspermobile/data/cache/report/InMemoryReportPageCache.java
/*
 * Copyright © 2016 TIBCO Software,Inc.All rights reserved.
 * http://community.jaspersoft.com/project/jaspermobile-android
 *
 * Unless you have purchased a commercial license agreement from TIBCO Jaspersoft,
 * the following license terms apply:
 *
 * This program is part of TIBCO Jaspersoft Mobile for Android.
 *
 * TIBCO Jaspersoft Mobile is free software:you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published by
 * the Free Software Foundation,either version 3of the License,or
 * (at your option)any later version.
 *
 * TIBCO Jaspersoft Mobile is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY;without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
 * GNU Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with TIBCO Jaspersoft Mobile for Android.If not,see
 * <http://www.gnu.org/licenses/lgpl>.
 */

package com.jaspersoft.android.jaspermobile.data.cache.report;

import android.support.annotation.NonNull;
import android.support.annotation.Nullable;
import android.support.v4.util.LruCache;

import com.jaspersoft.android.jaspermobile.domain.PageRequest;
import com.jaspersoft.android.jaspermobile.domain.ReportPage;
import com.jaspersoft.android.jaspermobile.internal.di.PerProfile;

import javax.inject.Inject;

/**
 * @author Tom Koptel
 * @since 2.3
 */
@PerProfile
public final class InMemoryReportPageCache implements ReportPageCache {
    private final LruCache<String, ReportPage> mStorage = new LruCache<>(10);

    @Inject
    public InMemoryReportPageCache() {
    }

    @Nullable
    @Override
    public ReportPage get(@NonNull PageRequest pageRequest) {
        String id = pageRequest.getIdentifier();
        return mStorage.get(id);
    }

    @Override
    @NonNull
    public ReportPage put(@NonNull PageRequest pageRequest, @NonNull ReportPage content) {
        String id = pageRequest.getIdentifier();
        mStorage.put(id, content);
        return content;
    }

    @Override
    public void evictAll() {
        mStorage.evictAll();
    }
}
